+++ src/pages/User/ui/TabSection/LookbookList/LookbookCard/LookbookCard.style.ts
