+++ src/pages/User/ui/TabSection/SettingList/ThicknessSetMenu/ThicknessSetMenu.tsx
