+++ src/pages/User/ui/TabSection/TabSection.style.ts
