+++ src/pages/User/ui/TabSection/TabSection.tsx
