+++ src/pages/User/ui/UserPage.tsx
@@ -1,9 +1,8 @@
 import { useAppSelector } from '@/shared/lib/useAppSelector';
 import { HeadHelmet } from '@/shared/ui';
 
-import TabSection from '../../TabSection/TabSection';
-import UserHeader from '../../UserHeader/UserHeader';
-
+import TabSection from './TabSection/TabSection';
+import UserHeader from './UserHeader/UserHeader';
 import { S } from './UserPage.style';
 
 export const UserPage = () => {
